demo-huggingface/requirements.txt
grad-cam==1.5.5
gradio==4.44.1
matplotlib==3.9.4
numpy==2.0.2
opencv-python==4.11.0.86
pillow==10.4.0
torch==2.7.1
torchvision==0.22.1
